+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## v1.3.0 - 2024-12-02
+
 ### Added
 
 - Define a `thumbnail` relationship on the post model that when defined will return an attachment model.
+- Added count support to `assertSee()` for responses and `assertContains()` for HTML string assertions.
+- Added `assertContains()` to response assertions.
 
 ### Changed
 
@@ -25,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ensure that `with_terms()` can support an array of term slugs when passed with a
   taxonomy index.
 - Ensure that framework configuration respects the application configuration.
+- Ensure that collections can properly implode `Stringable` objects.
 
 ## v1.2.0 - 2024-09-23